+++ src/Blitz.Interfacing/ExceptionResult.cs
@@ -6,13 +6,14 @@ using MessagePack;
 [MessagePackObject]
 public class ExceptionResult
 {
-    
+
     [Key(nameof(ExceptionMessage))]
-    public string? ExceptionMessage { get; set; }
+    public string ExceptionMessage { get; set; } = string.Empty;
 
 
-    [Key(nameof(ExceptionStack))] 
-    public string? ExceptionStack { get; set; }
+    [Key(nameof(ExceptionStack))]
+
+    public string ExceptionStack { get; set; } = string.Empty;
 
 
     public static ExceptionResult CreateFromException(Exception ex)